backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1

# Supabase
supabase==2.0.3
postgrest==0.12.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Caching
redis==5.0.1
aioredis==2.0.1

# Background Tasks
celery==5.3.4
celery[redis]==5.3.4

# Data Processing
numpy==1.25.2
pandas==2.1.4
scikit-learn==1.3.2

# Machine Learning (Optional)
tensorflow==2.15.0  # For advanced ML features
torch==2.1.2  # Alternative ML framework

# Text Processing
janome==0.5.0  # Japanese morphological analyzer
transformers==4.36.0  # NLP models

# File Processing
python-xlsx==0.2.4
openpyxl==3.1.2
python-docx==1.1.0

# Email
aiosmtplib==3.0.1
jinja2==3.1.2
premailer==3.10.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.39.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing
factory-boy==3.3.0  # Test data generation

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Utilities
python-dotenv==1.0.0
click==8.1.7
typer==0.9.0
rich==13.7.0
tqdm==4.66.1

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Image Processing (Optional)
Pillow==10.1.0

# Geographic Data (Optional)
geopy==2.4.1

# Configuration
pyyaml==6.0.1
toml==0.10.2

# Performance
ujson==5.8.0
orjson==3.9.10

# Memory Profiling (Development)
memory-profiler==0.61.0
psutil==5.9.6

# AWS SDK (if using AWS services)
boto3==1.34.0
botocore==1.34.0

# Google Cloud (if using GCP services)
google-cloud-storage==2.10.0
google-cloud-bigquery==3.13.0

# Stripe (if using payment processing)
stripe==7.8.0

# OpenAI (if using GPT features)
openai==1.3.7

# Slack (if using Slack notifications)
slack-sdk==3.26.0

# Export formats
xlsxwriter==3.1.9
reportlab==4.0.7  # PDF generation